refactor: :art: replace all file names to kebad case
