+++ package.json
@@ -45,7 +45,6 @@
     "@commitlint/config-conventional": "^16.2.1",
     "@commitlint/cz-commitlint": "^15.0.0",
     "@cypress/vite-dev-server": "^2.2.2",
-    "@eztek/prettier-config": "^1.1.0",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
