+++ src/features/auth/routes/auth-routes.tsx
@@ -1,8 +1,8 @@
 import { lazy } from 'react';
 import { RouteObject } from 'react-router-dom';
 
-const SignInPage = lazy(() => import('../views/SignInPage/SignInPage'));
-const SignUpPage = lazy(() => import('../views/SignUpPage/SignUpPage'));
+const SignInPage = lazy(() => import('../views/sign-in-page/sign-in-page'));
+const SignUpPage = lazy(() => import('../views/sign-up-page/sign-up-page'));
 
 const authRoutes: Array<RouteObject> = [
   {
