+++ src/features/dashboard/routes/index.ts
@@ -1,0 +1,1 @@
+export { default } from './dashboard-routes';
