+++ src/layout/minimal-layout/minimal-layout.tsx
