+++ src/layout/portal-layout/portal-layout.tsx
@@ -1,4 +1,4 @@
-import StackedLayout from './StackedLayout/StackedLayout';
+import StackedLayout from './stacked-layout/stacked-layout';
 
 export default function PortalLayout() {
   return <StackedLayout />;
