+++ src/main.tsx
@@ -4,7 +4,7 @@ import { StrictMode } from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import App from './App';
+import App from './app';
 
 const rootElement = document.getElementById('root');
 
@@ -14,5 +14,5 @@ ReactDOM.render(
       <App />
     </BrowserRouter>
   </StrictMode>,
-  rootElement
+  rootElement,
 );
