+++ src/routes/loader/index.ts
@@ -1,0 +1,1 @@
+export * from './loader';
