+++ src/routes/loader/loader.tsx
@@ -1,0 +1,3 @@
+export const Loader = () => {
+  return <div>Loading...</div>;
+};
