+++ src/ui/button/button.tsx
@@ -1,0 +1,3 @@
+export const Button = () => {
+  return <button>Button</button>;
+};
