+++ src/ui/dialog/dialog.tsx
@@ -1,0 +1,3 @@
+export const Dialog = () => {
+  return <div>Dialog</div>;
+};
